# Optimize buffer sizes for bytestream.Write