--- a/enterprise/server/util/distributed_client/distributed_client.go
+++ b/enterprise/server/util/distributed_client/distributed_client.go
@@ -33,6 +33,8 @@ const (
 	// Keep under the limit of ~4MB (save 256KB).
 	// (Match the readBufSizeBytes in byte_stream_server.go)
 	readBufSizeBytes = (1024 * 1024 * 4) - (1024 * 256)
+
+	writeBufSizeBytes = 512 * 1024 // 512 KiB
 )
 
 type Proxy struct {
@@ -688,8 +690,8 @@ func safeBufferSize(r *rspb.ResourceName) int {
 	if r.GetCacheType() == rspb.CacheType_CAS {
 		size = int(r.GetDigest().GetSizeBytes())
 	}
-	if size > readBufSizeBytes {
-		size = readBufSizeBytes
+	if size > writeBufSizeBytes {
+		size = writeBufSizeBytes
 	}
 	return size
 }
@@ -720,7 +722,7 @@ func (c *Proxy) RemoteWriter(ctx context.Context, peer, handoffPeer string, r *r
 		stream:      stream,
 		r:           r,
 	}
-	return ioutil.NewDoubleBufferWriter(ctx, wc, c.readBufPool, safeBufferSize(r), readBufSizeBytes), nil
+	return ioutil.NewDoubleBufferWriter(ctx, wc, c.readBufPool, safeBufferSize(r), writeBufSizeBytes), nil
 }
 
 func (c *Proxy) SendHeartbeat(ctx context.Context, peer string) error {